--- a/src/common/legacy_config_opts.h
+++ b/src/common/legacy_config_opts.h
@@ -1549,3 +1549,4 @@ OPTION(debug_allow_any_pool_priority, OPT_BOOL)
 OPTION(rgw_gc_max_deferred_entries_size, OPT_U64) // GC deferred entries size in queue head
 OPTION(rgw_gc_max_queue_size, OPT_U64) // GC max queue size
 OPTION(rgw_gc_max_deferred, OPT_U64) // GC max number of deferred entries
+OPTION(rgw_gc_defer_enabled, OPT_BOOL) // Turn on/off defer GC