… src/common/options.cc
… src/rgw/rgw_op.cc