@pritha-srivastava 2020-12-02
This comment is with respect to the last commit - I noticed that radosgw-admin gc list is slower than the --include-all version. While I was investigating the reason behind this I discovered that using in memory transitioned_objects_cache in RGWGC::list() doesn't help since for every radosgw-admin command a new RGWGC object is instantiated and hence a new transitioned_objects_cache is instantiated, and for every gc list call we make 2 cls_rgw_gc_list calls and for every gc list --include call we make one cls_rgw_gc_list call. Instead of using an in memory variable, I have replaced that with the version of gc object being set to 2 when the transition to queue happens and the version is checked each time a gc list and gc list --include call is made. If the version of the gc object (queue) has already been set to 2 by RGWGC::process, then the cls_rgw_gc_list calls are skipped. If it hasn't been changed to 2, then the first gc list or gc list --include-all call will change it to 2, and subsequent gc list calls won't issue cls_rgw_gc_list calls. I can see a slight difference in performance when I delete 50 objects in my system, although I don't know what the impact will be on a larger scale.
Before the change: gc list : 0.16 seconds gc list --include-all: 0.12 seconds
After the change: gc list: 0.08 seconds gc list --include-all: 0.08 seconds