@pritha-srivastava 2020-12-03
> I'm unclear if this is just performance fixes or if this also fixes the underlying gc defer bug too.

I asked Casey @cbodley  on the other PR if I can add perf fixes to this in addition to gc defer bug fix, and I think he agreed (he gave a thumbs up), in case we do not want to club them together, I can create a separate PR for that.