@pritha-srivastava 2021-01-18
Teuthology run results here: https://pulpito.ceph.com/prsrivas-2021-01-15_10:43:52-rgw-wip-rgw-gc-defer-perf-fixes-distro-basic-smithi/, the failures are not related to this PR.